[CrashMonitor] Create a UI with MonoMac and reorganise the code a bit
The UI for the CrashMonitor is now written using monomac to simplify
integrating nicely with macos. We can now easily hide the dock icon
and ensure the error messages show up correctly onscreen.

Integrate the building of this with the regular monodevelop build.
Everything should be built when you hit 'make' now.